re-enable semistandard checking on CodeShip. https://github.com/numo-labs/lambda-taggable-createDocument/issues/16 should be ready!
