+++ lib/geonames_cache.js
@@ -1,7 +1,6 @@
 var AWS = require('aws-sdk');
 AWS.config.region = 'eu-west-1';
 var s3 = new AWS.S3();
-// var s3bucket = new AWS.S3({params: {Bucket: }});
 
 module.exports.set = function set (path, record, callback) {
   var filepath = path + '.json';
@@ -14,32 +13,31 @@ module.exports.set = function set (path, record, callback) {
     ACL: 'public-read'
   };
   s3.upload(params, function (err, data) {
-		record.Key = data.Key;
-		return callback(err, record);
+    record.Key = data.Key; // return the original record with S3 Key (Mutation!)
+    return callback(err, record);
   });
 };
 
 // get from cache
 module.exports.get = function get (path, callback) {
-	var params = {
-		Bucket: process.env.GEONAMES_CACHE, /* required */
-		Key: path + '.json'
-	};
-	s3.getObject(params, function(err, data) {
-		// console.log(err, data);
-		if(!err && data.Body) {
-	  	return callback(err, JSON.parse(data.Body));
-		} else {
-			return callback(err, data);
-		}
-	});
+  var params = {
+    Bucket: process.env.GEONAMES_CACHE, /* required */
+    Key: path + '.json'
+  };
+  s3.getObject(params, function (err, data) {
+  // console.log(err, data);
+    if (!err && data.Body) {
+      return callback(err, JSON.parse(data.Body));
+    } else {
+      return callback(err, data);
+    }
+  });
 };
 
 module.exports.remove = function remove (path, callback) {
-
-	var params = {
-		Bucket: process.env.GEONAMES_CACHE, /* required */
-		Key: path + '.json'
-	};
-	s3.deleteObject(params, callback);
-}
+  var params = {
+    Bucket: process.env.GEONAMES_CACHE, /* required */
+    Key: path + '.json'
+  };
+  s3.deleteObject(params, callback);
+};
